+++ docker/entrypoint.sh
@@ -217,7 +217,9 @@ websockify --web="$NOVNC_ROOT" "$WEB_PORT" "localhost:$VNC_PORT" \
 WEB_PID=$!
 
 log ""
-log "  play at  http://localhost:$WEB_PORT/vnc.html?autoconnect=1&resize=off"
+log "  play at  http://localhost:$WEB_PORT/play.html"
+log ""
+log "  plain noVNC (no mouse capture):  http://localhost:$WEB_PORT/vnc.html?autoconnect=1&resize=off"
 log ""
 
 ##############################################################################